fix ambiguous truth value of an array
Use "if len(artifact_times)" instead of just "if artifact_times" because truth value of an array is ambiguous. The latter gave me an error about the ambiguity of "if artifact_times" when there was only one artifact detected. Can someone please review whether this is the intended boolean evaluation to use here, or whether another use such as .any() or .all() is what was intended? Thanks.