Enable adding >1 dataframes to an analysis nwb
See issue #252 . This change enables the user to set the name of the dynamic table when created from a dataframe before adding it as an nwb object to an analysis nwb file. Previously, the name was hard coded as 'pandas_table' which disallowed the entry of more than one dataframe into an analysis nwb file, because they would have the same name. Now, the user can set a name for each table so that multiple dataframes can be convereted to hdmf dynamic tables and added as unique nwb objects to an analysis nwb file. Tested locally. This change does not impact file reading as far as I understand, because nwb objects are fetched using the object id regardless of the nwb object type.